+++ src/main/kotlin/io/reactivex/rxkotlin/maybe.kt
@@ -1,4 +1,4 @@
-package rx.lang.kotlin
+package io.reactivex.rxkotlin
 
 import io.reactivex.Maybe
 import java.util.concurrent.Callable
