+++ src/main/kotlin/io/reactivex/rxkotlin/single.kt
@@ -1,4 +1,4 @@
-package rx.lang.kotlin
+package io.reactivex.rxkotlin
 
 import io.reactivex.Single
 import io.reactivex.SingleEmitter
