+++ src/test/kotlin/io/reactivex/rxkotlin/SingleTest.kt
@@ -1,4 +1,4 @@
-package rx.lang.kotlin
+package io.reactivex.rxkotlin
 
 import io.reactivex.Observable
 import io.reactivex.Single
