+++ src/test/kotlin/io/reactivex/rxkotlin/SubscriptionTests.kt
@@ -1,4 +1,4 @@
-package rx.lang.kotlin
+package io.reactivex.rxkotlin
 
 import io.reactivex.Observable
 import io.reactivex.disposables.CompositeDisposable
